@@ -4,20 +4,20 @@ publish_to: 'none'
 version: 0.1.0
 
 environment:
-  sdk: '>=3.4.4 <4.0.0'
+  sdk: '>=3.5.0 <4.0.0'
 
 dependencies:
-  async: ^2.11.0
-  country_flags: ^3.0.0
-  equatable: ^2.0.5
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
+  async: ^2.11.0
+  country_flags: ^3.0.0
+  equatable: ^2.0.5
   flutter_svg: ^2.0.10+1
-  hive_flutter: ^1.1.0
   image_picker: ^1.1.2
   localization: ^2.1.1
+  path_provider: ^2.1.4
   photo_view: ^0.15.0
   shared_preferences: ^2.3.1
   uuid: ^4.4.2
@@ -26,10 +26,8 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  analyzer: ^4.7.0
-  flutter_lints: ^3.0.0
-  hive_generator: ^1.1.2
-  build_runner: ^2.1.8
+  analyzer: ^5.13.0
+  flutter_lints: ^4.0.0
 
 flutter:
   uses-material-design: true
