@@ -1,25 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:hive_flutter/hive_flutter.dart';
 import 'package:localization/localization.dart';
 import 'package:my_fit_journey/data.dart';
-import 'package:my_fit_journey/models/body_part.dart';
 import 'package:my_fit_journey/models/exercise.dart';
-import 'package:my_fit_journey/models/exercise_photo.dart';
 import 'package:my_fit_journey/pages/exercise_page.dart';
 import 'package:my_fit_journey/pages/body_selector_page.dart';
 import 'package:my_fit_journey/pages/main_page.dart';
+import 'package:my_fit_journey/storage/storage.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
-  await Hive.initFlutter();
-
-  Hive.registerAdapter(ExerciseAdapter());
-  Hive.registerAdapter(ExercisePhotoAdapter());
-  Hive.registerAdapter(BodyPartAdapter());
-
   await initData();
+  await Storage.initialize();
 
   runApp(const MainApp());
 }
